pyproject.toml
[tool.poetry]
name = "extract-from-url"
version = "0.3.0"
description = "Download and extract files on-the-fly (ZIP files too)"
authors = ["Francesco Frassinelli <francesco.frassinelli@nina.no>"]
license = "GPL-3.0+"
readme = "README.rst"
homepage = "https://github.com/frafra/extract-from-url"
repository = "https://github.com/frafra/extract-from-url"
keywords = ["libarchive-c"]

[tool.poetry.dependencies]
python = "^3.7"
libarchive-c = ">=1.0"
fs = ">=2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
